created separate function for cover art
