+++ static/img/credits.txt
@@ -1,5 +1,26 @@
-underwaterBubbles.jpg photo by Cristian Palmer on Unsplash:
+Album cover art:
+electro-music-album_03.png by Designed by rawpixel.com
+https://www.freepik.com/free-vector/electro-music-album_3601853.htm
+
+electro-music-album_02.png by Designed by rawpixel.com
+https://www.freepik.com/free-vector/electro-music-album_3601859.htm
+
+electro-music-album_01.png by Designed by rawpixel.com
+https://www.freepik.com/free-vector/electro-music-album_3601857.htm
+
+electro-music-album_00.png by Designed by rawpixel.com
+https://www.freepik.com/free-vector/electro-music-album_3601855.htm
+
+cd-cover-blank.png by Designed by rawpixel.com
+https://www.freepik.com/free-vector/cd-cover-design-mockup_3463208.htm
+
+
+Background image art:
+underwater-bubbles.jpg photo by Cristian Palmer on Unsplash:
 https://unsplash.com/photos/3leBubkp5hk
 
-blankAlbum.jpg by Designed by rawpixel.com
-https://www.freepik.com/free-vector/electro-music-album_3601853.htm
+clear-blue.jpeg by Sven Scheuermeier, https://unsplash.com/@sveninho
+https://unsplash.com/photos/saGbrA6s8g0
+
+taking-dip.jpeg by Tim Marshall, https://unsplash.com/@timmarshall
+https://unsplash.com/photos/T-AOFf2B8Zw
